Revert Scheme: get-user-name/get-user-login from Max

+++ src/Plugins/Qt/qt_gui.cpp
@@ -878,7 +878,7 @@ qt_gui_rep::update () {
   
   time_t delay = delayed_commands.lapse - texmacs_time();
   if (needing_update) delay = 0;
-  else                delay = max (0, min (std_delay, static_cast<int>(delay)));
+  else                delay = max (0, min (std_delay, delay));
   if (postpone_treatment) delay= 9; // NOTE: force occasional display
  
   updatetimer->start (delay);
